Separate vue examples
