+++ examples/angular-cli/.storybook/config.js
@@ -13,8 +13,8 @@ function loadStories() {
   require('../src/stories');
 
   // automatically import all story ts files that end with *.stories.ts
-  const req = require.context('../src/stories', true, /\.stories\.ts$/)
-  req.keys().forEach((filename) => req(filename))
+  const req = require.context('../src/stories', true, /\.stories\.ts$/);
+  req.keys().forEach((filename) => req(filename));
 }
 
 configure(loadStories, module);
