+++ examples/vue-kitchen-sink/.storybook/addons.js
@@ -3,3 +3,4 @@ import '@storybook/addon-links/register';
 import '@storybook/addon-notes/register';
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-viewport/register';
+import '@storybook/addon-options/register';
