+++ examples/vue-kitchen-sink/src/stories/addon-actions.stories.js
@@ -1,0 +1,19 @@
+import { storiesOf } from '@storybook/vue';
+import { action } from '@storybook/addon-actions';
+
+storiesOf('Addon|Actions', module)
+  .add('Action only', () => ({
+    template: '<my-button :handle-click="log">Click me to log the action</my-button>',
+    methods: {
+      log: action('log1'),
+    },
+  }))
+  .add('Action and method', () => ({
+    template: '<my-button :handle-click="log">Click me to log the action</my-button>',
+    methods: {
+      log: e => {
+        e.preventDefault();
+        action('log2')(e.target);
+      },
+    },
+  }));
