+++ examples/vue-kitchen-sink/src/stories/addon-centered.stories.js
@@ -1,0 +1,11 @@
+import { storiesOf } from '@storybook/vue';
+import Centered from '@storybook/addon-centered';
+
+import MyButton from './Button.vue';
+
+storiesOf('Addon|Centered', module)
+  .addDecorator(Centered)
+  .add('rounded', () => ({
+    components: { MyButton },
+    template: '<my-button :rounded="true">A Button with rounded edges</my-button>',
+  }));
